refactor: split flaky concurrent report test into own target
Extract 'creating 3 reports at once succeeds' from
InProcessDirectOnlyReportIntegrationTest into a new
InProcessDirectMultipleReportIntegrationTest class with its own
GCloud test target marked flaky = True. This isolates the known
concurrent BatchCreateMetrics race condition from the stable
direct-only tests.
