+++ src/main/kotlin/org/wfanet/measurement/integration/common/reporting/v2/BUILD.bazel
@@ -11,6 +11,7 @@ package(
 kt_jvm_library(
     name = "in_process_life_of_a_report_integration_test",
     srcs = [
+        "InProcessDirectMultipleReportIntegrationTest.kt",
         "InProcessDirectOnlyReportIntegrationTest.kt",
         "InProcessLifeOfAReportIntegrationTest.kt",
         "InProcessMultiEdpReportIntegrationTest.kt",
