+++ src/main/kotlin/org/wfanet/measurement/integration/common/reporting/v2/InProcessDirectMultipleReportIntegrationTest.kt
@@ -1,0 +1,158 @@
+/*
+ * Copyright 2026 The Cross-Media Measurement Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.wfanet.measurement.integration.common.reporting.v2
+
+import com.google.common.truth.Truth.assertThat
+import com.google.common.truth.extensions.proto.ProtoTruth.assertThat
+import com.google.protobuf.timestamp
+import com.google.type.interval
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.async
+import kotlinx.coroutines.awaitAll
+import kotlinx.coroutines.runBlocking
+import org.junit.Test
+import org.wfanet.measurement.api.v2alpha.event_templates.testing.Person
+import org.wfanet.measurement.common.testing.ProviderRule
+import org.wfanet.measurement.integration.common.ALL_DUCHY_NAMES
+import org.wfanet.measurement.integration.common.AccessServicesFactory
+import org.wfanet.measurement.integration.common.InProcessDuchy
+import org.wfanet.measurement.kingdom.deploy.common.service.DataServices
+import org.wfanet.measurement.reporting.deploy.v2.common.service.Services
+import org.wfanet.measurement.reporting.v2alpha.EventGroup
+import org.wfanet.measurement.reporting.v2alpha.MetricSpecKt
+import org.wfanet.measurement.reporting.v2alpha.Report
+import org.wfanet.measurement.reporting.v2alpha.ReportKt
+import org.wfanet.measurement.reporting.v2alpha.ReportingSet
+import org.wfanet.measurement.reporting.v2alpha.createMetricCalculationSpecRequest
+import org.wfanet.measurement.reporting.v2alpha.createReportRequest
+import org.wfanet.measurement.reporting.v2alpha.listReportsRequest
+import org.wfanet.measurement.reporting.v2alpha.metricCalculationSpec
+import org.wfanet.measurement.reporting.v2alpha.metricSpec
+import org.wfanet.measurement.reporting.v2alpha.report
+import org.wfanet.measurement.reporting.v2alpha.timeIntervals
+import org.wfanet.measurement.system.v1alpha.ComputationLogEntriesGrpcKt
+
+abstract class InProcessDirectMultipleReportIntegrationTest(
+  kingdomDataServicesRule: ProviderRule<DataServices>,
+  duchyDependenciesRule:
+    ProviderRule<
+      (
+        String, ComputationLogEntriesGrpcKt.ComputationLogEntriesCoroutineStub,
+      ) -> InProcessDuchy.DuchyDependencies
+    >,
+  accessServicesFactory: AccessServicesFactory,
+  reportingDataServicesProviderRule: ProviderRule<Services>,
+  duchyNames: List<String> = ALL_DUCHY_NAMES,
+  hmssEnabled: Boolean,
+  trusTeeEnabled: Boolean,
+) :
+  InProcessLifeOfAReportIntegrationTest(
+    kingdomDataServicesRule,
+    duchyDependenciesRule,
+    accessServicesFactory,
+    reportingDataServicesProviderRule,
+    duchyNames,
+    hmssEnabled,
+    trusTeeEnabled,
+  ) {
+
+  @Test
+  fun `creating 3 reports at once succeeds`() = runBlocking {
+    val numReports = 3
+    val measurementConsumerData = inProcessCmmsComponents.getMeasurementConsumerData()
+    val eventGroups = listEventGroups()
+    val eventGroupEntries: List<Pair<EventGroup, String>> =
+      listOf(eventGroups.first() to "person.age_group == ${Person.AgeGroup.YEARS_18_TO_34_VALUE}")
+    val createdPrimitiveReportingSet: ReportingSet =
+      createPrimitiveReportingSets(eventGroupEntries, measurementConsumerData.name).single()
+
+    val createdMetricCalculationSpec =
+      publicMetricCalculationSpecsClient
+        .withCallCredentials(credentials)
+        .createMetricCalculationSpec(
+          createMetricCalculationSpecRequest {
+            parent = measurementConsumerData.name
+            metricCalculationSpec = metricCalculationSpec {
+              displayName = "load test"
+              metricSpecs += metricSpec {
+                reach = MetricSpecKt.reachParams { privacyParams = DP_PARAMS }
+                vidSamplingInterval = VID_SAMPLING_INTERVAL
+              }
+            }
+            metricCalculationSpecId = "fed"
+          }
+        )
+
+    val report = report {
+      reportingMetricEntries +=
+        ReportKt.reportingMetricEntry {
+          key = createdPrimitiveReportingSet.name
+          value =
+            ReportKt.reportingMetricCalculationSpec {
+              metricCalculationSpecs += createdMetricCalculationSpec.name
+            }
+        }
+      timeIntervals = timeIntervals {
+        timeIntervals += interval {
+          startTime = timestamp { seconds = 100 }
+          endTime = timestamp { seconds = 200 }
+        }
+      }
+    }
+
+    val deferred: MutableList<Deferred<Report>> = mutableListOf()
+    repeat(numReports) {
+      deferred.add(
+        async {
+          publicReportsClient
+            .withCallCredentials(credentials)
+            .createReport(
+              createReportRequest {
+                parent = measurementConsumerData.name
+                this.report = report
+                reportId = "report$$it"
+              }
+            )
+        }
+      )
+    }
+
+    deferred.awaitAll()
+    val retrievedReports =
+      publicReportsClient
+        .withCallCredentials(credentials)
+        .listReports(
+          listReportsRequest {
+            parent = measurementConsumerData.name
+            pageSize = numReports
+          }
+        )
+        .reportsList
+
+    assertThat(retrievedReports).hasSize(numReports)
+    retrievedReports.forEach {
+      assertThat(it)
+        .ignoringFields(
+          Report.NAME_FIELD_NUMBER,
+          Report.STATE_FIELD_NUMBER,
+          Report.CREATE_TIME_FIELD_NUMBER,
+          Report.METRIC_CALCULATION_RESULTS_FIELD_NUMBER,
+        )
+        .isEqualTo(report)
+    }
+  }
+}
